+++ ichnaea/models/constants.py
@@ -52,6 +52,8 @@ as a test mac address and ignore data submissions with it.
 INVALID_MAC_REGEX = re.compile('(?!(0{12}|f{12}|%s))' % WIFI_TEST_MAC)
 VALID_MAC_REGEX = re.compile('([0-9a-fA-F]{12})')
 
+VALID_APIKEY_REGEX = re.compile('^[-0-9a-z]+$', re.IGNORECASE | re.UNICODE)
+
 MIN_WIFI_CHANNEL = 0  #: Minimum accepted WiFi channel.
 MAX_WIFI_CHANNEL = 166  #: Maximum accepted WiFi channel.