+++ CHANGES.rst
@@ -14,6 +14,8 @@ Migrations
 Changes
 ~~~~~~~
 
+- Restrict valid characters in API keys.
+
 - Take signal strength into account for location queries.
 
 - Decrease database session times in data tasks.